Fix dangerous PATH definition in cron
This PATH definition based on current ::path fact breaks when puppet
is run in the same crontab without its own PATH definition. Each puppet
run adds one `/opt/puppet/bin` to the variable.

The `::path` fact should probably never be used in a Puppet recipe as it
can change depending on the environment in which puppet is run.